requirements.txt
ollama>=0.1.0
pyyaml>=6.0
pytest>=7.4.0
pylint>=3.0.0
black>=23.0.0
streamlit>=1.28.0